@@ -1,10 +1,11 @@
-const mongoose = require('mongoose')
-    , dbUrl = process.env.MONGODB_URI || 'mongodb://localhost:27017/database'
+import mongoose from 'mongoose'
 
-module.exports = (url = dbUrl) => {
+const dbUrl = process.env.MONGODB_URI || 'mongodb://localhost:27017/database'
+
+export default (url = dbUrl) => {
   mongoose.Promise = global.Promise
-  mongoose.connect(url)
+  mongoose.connect(url, { useMongoClient: true })
   mongoose.connection.on('error', (err) => {
-      console.log('Mongoose Connection ERROR: ' + err)
+    console.log(`Mongoose Connection ERROR: ${err}`)
   })
 }
