@@ -1,5 +1,6 @@
-const mongoose = require('mongoose')
-    , { Model, Schema } = mongoose
+import mongoose from 'mongoose'
+
+const { Model, Schema } = mongoose
 
 const schema = new Schema(
   { id: Schema.ObjectId
@@ -10,4 +11,4 @@ const schema = new Schema(
 
 class genericModel extends Model { }
 
-module.exports = mongoose.model(genericModel, schema, 'collection')
+export default mongoose.model(genericModel, schema, 'collection')
