@@ -1,22 +1,25 @@
+import cache from 'express-cache-headers'
+import express from 'express'
+import logger from 'morgan'
+import path from 'path'
 
-let express = require('express')
-  , logger = require('morgan')
-  , app = express()
-  , addRoutes = require('./routes')
-  , attachSocket = require('./sockets')
-  , port = process.env.PORT || 3000
-  , path = require('path')
-  , cache = require('express-cache-headers')
-  , connectDatabase = require('./database.js')
+import routes from './routes/routes'
+import apiRoutes from './routes/api-routes'
+import attachSocket from './sockets'
+import connectDatabase from './database'
+
+const port = process.env.PORT || 3000
+
+let app = express()
 
 app.use(cache(60))
 app.use(logger('dev'))
 app.use(express.static(path.join(__dirname, 'assets', 'build')))
-addRoutes(app)
-
+apiRoutes(app)
+routes(app)
 connectDatabase()
 
 let server = app.listen(port, () => {
-  console.log(`Listening on http://localhost:${port}`)
   attachSocket(server)
+  console.log(`Listening on http://localhost:${port}`)
 })
