@@ -1,28 +1,28 @@
-const Socket = require('socket.io')
+import Socket from 'socket.io'
 
-module.exports = (server) => {
+export default (server) => {
   const io = Socket.listen(server)
   const ROOM_SIZE = 2
 
   io.sockets.on('connection', (socket) => {
 
-    console.log(socket.id + ': client connected')
+    console.log(`${socket.id}: client connected`)
 
     socket.on('error', console.log)
 
-    socket.on('join', function (room) {
+    socket.on('join', (room) => {
       if (io.sockets.adapter.rooms[room] && io.sockets.adapter.rooms[room].length < ROOM_SIZE) {
         socket.join(room)
         socket.room = room
-        io.sockets.in(room).emit('joined', socket.id + ' has joined')
+        io.sockets.in(room).emit('joined', `${socket.id} has joined`)
       } else {
         socket.emit('joined', 'Room full :(')
-       }
+      }
     })
 
     socket.on('disconnect', () => {
-      io.sockets.in(socket.room).emit('left', socket.id + ' has left')
-      console.log(socket.id + ': client disconnected')
+      io.sockets.in(socket.room).emit('left', `${socket.id} has left`)
+      console.log(`${socket.id}: client disconnected`)
     })
   })
 }
