@@ -1,4 +1,4 @@
-const assert = require('assert')
+import assert from 'assert'
 
 describe('Dummy class', () => {
 
@@ -11,7 +11,7 @@ describe('Dummy class', () => {
   })
 
   describe('dummy test', () => {
-    it('should pass!', function (done) {
+    it('should pass!', (done) => {
       assert.equal(true, true, 'Your tests are written badly!')
       done()
     })
